# perf: VID ADVZ verify_share use parallelism over multiplicity