--- a/vid/benches/advz_multiplicity.rs
+++ b/vid/benches/advz_multiplicity.rs
@@ -32,15 +32,15 @@ where
 {
     // play with these items
     //
-    // CODE_RATE is merely a convenient way to automatically choose recovery
-    // threshold as a function of storage node count. If desired, you could set
-    // recovery thresholds independent of storage node counts.
+    // INVERSE_CODE_RATE is merely a convenient way to automatically choose
+    // recovery threshold as a function of storage node count. If desired, you
+    // could set recovery thresholds independent of storage node counts.
     let multiplicities = [1, 256];
     let num_storage_nodes = 128;
-    const CODE_RATE: usize = 4; // ratio of num_storage_nodes : recovery_threshold
+    const INVERSE_CODE_RATE: usize = 4; // ratio of num_storage_nodes : recovery_threshold
 
     // more items as a function of the above
-    let recovery_threshold = num_storage_nodes / CODE_RATE;
+    let recovery_threshold = num_storage_nodes / INVERSE_CODE_RATE;
     let max_multiplicity = multiplicities.iter().max().unwrap();
     let max_degree = recovery_threshold * max_multiplicity;
     let coeff_byte_len = field_byte_len::<E::ScalarField>();