--- a/vid/src/advz.rs
+++ b/vid/src/advz.rs
@@ -139,8 +139,6 @@ where
         multiplicity: u32,       // batch m chunks, keep the rate r = (m*k)/(m*n)
         srs: impl Borrow<KzgSrs<E>>,
     ) -> VidResult<Self> {
-        // TODO support any degree, give multiple shares to nodes if needed
-        // https://github.com/EspressoSystems/jellyfish/issues/393
         if num_storage_nodes < recovery_threshold {
             return Err(VidError::Argument(format!(
                 "recovery_threshold {} exceeds num_storage_nodes {}",