… vid/.gitignore