@ggutoski 2024-08-13
> nit: 2 VID tests in `advz::bytes_to_field` are not working. Try `cargo test -p jf-vid --features test-srs --release -- --ignored` for details.

These doctests are ignored intentionally due to a Rust limitation that disallows access to private identifiers in doctest: [Doctest of private functions - help - The Rust Programming Language Forum](https://users.rust-lang.org/t/doctest-of-private-functions/60805) 😕