@ggutoski 2024-08-13
> And what's up with all the clippy warnings(?)

Github CI is using a newer version of Rust than our Nix environment. We need to update Nix flake again. We'll do that in another PR.